flask-app/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.0.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.2
protobuf==3.20.3
six==1.16.0
Werkzeug==2.2.3
cassandra-driver==3.25.0
python-arango


